Fix 'exec' builds on windows
