+++ cmd/exec.go
@@ -7,8 +7,6 @@ import (
     "github.com/spf13/cobra"
     "io"
     "os"
-    "os/signal"
-    "syscall"
 
     "k8s.io/client-go/kubernetes/scheme"
     "k8s.io/client-go/tools/remotecommand"
@@ -74,20 +72,6 @@ func (t *terminalSizeQueue) Next() *remotecommand.TerminalSize {
         return nil
     }
     return &size
-}
-
-func monitorTerminalResize(sizeQueue chan remotecommand.TerminalSize) {
-    ch := make(chan os.Signal, 1)
-    signal.Notify(ch, syscall.SIGWINCH)
-    defer signal.Stop(ch)
-
-    for range ch {
-        if width, height, err := term.GetSize(int(os.Stdin.Fd())); err == nil {
-            sizeQueue <- remotecommand.TerminalSize{
-                Width: uint16(width), Height: uint16(height),
-            }
-        }
-    }
 }
 
 func (sv *dshCmd) execPod(
